Cambios del home para hacerlo mas similar al figma
