+++ components/Toolbar/Toolbar.js
@@ -1,27 +1,72 @@
 import React from 'react'
-import Image from 'next/image'
-import ImgHero from '../Toolbar/img.png'
 import styles from '../../styles/Home.module.css'
-import Loguito from '../../public/images/paydeceLoguito.png'
+import SearchIcon from '@mui/icons-material/Search';
+import { styled, alpha } from '@mui/material/styles';
+import InputBase from '@mui/material/InputBase';
+
+
 
 function Toolbar() {
+
+  const Search = styled('div')(({ theme }) => ({
+    position: 'relative',
+    borderRadius: theme.shape.borderRadius,
+    backgroundColor: alpha(theme.palette.common.black, 0.90),
+    '&:hover': {
+      backgroundColor: alpha(theme.palette.common.black, 0.25),
+    },
+    marginRight: theme.spacing(2),
+    marginLeft: 0,
+    width: '100%',
+    [theme.breakpoints.up('sm')]: {
+      marginLeft: theme.spacing(3),
+      width: 'auto',
+    },
+  }));
+  
+  const SearchIconWrapper = styled('div')(({ theme }) => ({
+    padding: theme.spacing(0, 2),
+    height: '100%',
+    position: 'absolute',
+    pointerEvents: 'none',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  }));
+  
+  const StyledInputBase = styled(InputBase)(({ theme }) => ({
+    color: 'inherit',
+    '& .MuiInputBase-input': {
+      padding: theme.spacing(1, 1, 1, 0),
+      // vertical padding + font size from searchIcon
+      paddingLeft: `calc(1em + ${theme.spacing(4)})`,
+      transition: theme.transitions.create('width'),
+      width: '100%',
+      [theme.breakpoints.up('md')]: {
+        width: '50ch',
+      },
+    },
+  }));
+
     return(
-    <div className={styles.toolbar}/** */>
-        <Image className={ImgHero} src={ImgHero} alt="Img" />
+    <div className={styles.toolbar}>
         <div className={styles.hero}>
-        <h1 className={styles.title}>Bienvenido a <a target="_blank" rel="noopener noreferrer" href="https://www.instagram.com/criptocars/">criptocars</a></h1>
+        <h1 className={styles.title}>Comprá tu auto de forma segura con criptomonedas.</h1>
         
-        <p className={styles.description}>Comprá de forma segura con criptomonedas.</p>
-        <p className={styles.description}>
-          Utilizamos {'     '}
-          <a
-          href="https://paydece.io/"
-          target="_blank"
-          rel="noopener noreferrer"
-          >
-          <Image src={Loguito} alt="Logo" sx={{ display: { xs: 'none', md: 'flex' }, mr: 1 }}/>
-          </a> {'     '} como medio de pago, para brindarte una mayor seguridad.
-        </p>
+        {/* SearchBar */}
+        <Search>
+            <SearchIconWrapper>
+              <SearchIcon />
+            </SearchIconWrapper>
+            <StyledInputBase
+              placeholder="Escribí marca o modelo del vehículo que buscas..."
+              inputProps={{ 'aria-label': 'search' }}
+            />
+          </Search>
+          
+          <p className={styles.description}>
+            Utilizamos como medio de pago, para brindarte una mayor seguridad el Escrow de paydece.io
+          </p>
         </div>
     </div>
     );
